Moves ShareBtn into its own component
